+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 Update bugsnag-android to [v6.12.0](https://github.com/bugsnag/bugsnag-android/releases/tag/v6.12.0) [#882](https://github.com/bugsnag/bugsnag-unity/pull/882)
 Update bugsnag-cocoa to [v6.32.1](https://github.com/bugsnag/bugsnag-cocoa/releases/tag/v6.32.1) [#886](https://github.com/bugsnag/bugsnag-unity/pull/886)
+Update bugsnag-android to [v6.12.1](https://github.com/bugsnag/bugsnag-android/releases/tag/v6.12.1) [#888](https://github.com/bugsnag/bugsnag-unity/pull/888)
 
 ## 8.5.0 (2025-01-27)